OPlusExtras: AutoHBMService: Use ExecutorService for listener reg
Replicate what Google did for SystemUI in this commit.
aosp-mirror/platform_frameworks_base@fabc743

Registering a sensor seems to be an expensive operation,
and we do it on each screen-on event, so moving it to
an asynchronous task looks like a good idea anyway.

By moving all non-essential binder calls of the main thread or to the
next frame, we bring this down to 5ms, such that window animation
and Keyguard animation starts about at the same time.

The interesting part about the ExecutorService:
"Memory consistency effects: Actions in a thread prior to the submission
of a Runnable or Callable task to an ExecutorService happen-before any
actions taken by that task, which in turn happen-before the result is
retrieved via Future.get()."
(from https://developer.android.com/reference/java/util/concurrent/ExecutorService)

Change-Id: I4f37bb9a7dc9d7775d587d4ebd4b6619f3b77e81
Signed-off-by: AnierinB <anierin@evolution-x.org>
Signed-off-by: SathamHussainM <sathamhussain.m11@gmail.com>